Rename AndroidModule to AndroidClientModule Add DI config to AndroidClientModule Remove ICreateProfilePresenter as there is only one instance and not expected a use case for multiple instances. Add UserProfileRepository holding model and serviceFacade Adjust Screen and Presenter.
Discussion:
To me the usage of the repository does not make much sense here.
The 2 relevant methods do not map into the CRUD pattern which seems to be the design philosophy for repositories.
I would suggest to use it only when teh CRUD pattern and persistence is used and appropriate. And to inject the model and the service facade to the presenter in the other cases.
Otherwise we just create a useless wrapper.
@rodvar: If you would implement the usage here in a different way, can you provide with the given use case an implementation example?